requirements.txt
Pillow==7.1.2
boto3==1.13.10
